Update method stub to work with Rails 5
Haven't been able to find the change that caused this but before Rails 5
the relational object e.g. the event from `info_request.last_event` that
was returned inside the spec and the one used running code - in this
case in the view - had the same object reference. This is no longer the
case so any directly created method stub is ignored.